[PC-179] refactor: api 대신 presentation 이름 적용
