+++ api/src/main/java/org/yapp/domain/profile/presentation/response/ProfileValueResponse.java
@@ -1,4 +1,4 @@
-package org.yapp.domain.profile.api.response;
+package org.yapp.domain.profile.presentation.response;
 
 import org.yapp.domain.profile.ProfileValue;
